Consolidate into tango.core
tango.signal -> tango.core
tango.base_devices -> tango.core
tango -> tango.core
