--- src/ophyd_async/tango/base_devices/__init__.py
@@ -1,4 +1,0 @@
-from ._base_device import TangoDevice, tango_polling
-from ._tango_readable import TangoReadable
-
-__all__ = ["TangoDevice", "TangoReadable", "tango_polling"]
