+++ src/ophyd_async/tango/core/_signal.py
